Fix webhook tests and issues found during fixing them
